better scripts

+++ ingress/de-activate.sh
@@ -1,25 +1,33 @@
-cd pepsi
+#!/bin/bash
 
-echo "Delete Pepsi Resources"
+delete_brand_resources() {
+    local brand_name=$1
 
-kubectl delete -f ./ 2>/dev/null
+    echo " "
+    echo "Deleting ${brand_name} Resources"
 
-if [ $? -ne 0 ]; then
-    echo "There are no Pepsi resources currently up"
+    kubectl delete -f  ./  2>/dev/null
+
+    if [ $? -ne 0 ]; then
+        echo "There are no ${brand_name} resources currently up"
+    fi
+}
+
+
+read -p "This script will delete resources for Pepsi and Coca-Cola. Are you sure you want to continue? (y/n): " confirm
+if [ "$confirm" != "y" ]; then
+    echo "Aborting."
+    exit 0
 fi
 
+cd pepsi
+delete_brand_resources "Pepsi"
 cd ..
 
 cd cocacola
-
-echo "Delete Coca-Cola Resources"
-
-kubectl delete -f ./ 2>/dev/null
-
-if [ $? -ne 0 ]; then
-    echo "There are no Coca-Cola resources currently up"
-fi
-
+delete_brand_resources "Coca-Cola"
 cd ..
 
+echo " "
+echo "Deleting ingress"
 kubectl delete -f drinks-ingress.yml
